hobot/requirements.txt
python-kis
python-dotenv
openai
bs4
langgraph
langchain
langchain-openai
jupyter
langchain-community

schedule
pyupbit
slack_sdk
fastapi
uvicorn[standard]
python-multipart
gunicorn
PyJWT
email-validator

langchain_google_genai
pytz
google
google-genai

langchain-opentutorial

langgraph
langsmith
langchain-tavily

# MySQL
pymysql

# Encryption
cryptography

# MFA (Multi-Factor Authentication)
pyotp
qrcode
Pillow

# Macro Trading
fredapi
pandas
numpy
requests
yfinance
beautifulsoup4
lxml
selenium  # JavaScript 렌더링이 필요한 웹페이지 크롤링용 (선택적)
webdriver-manager  # ChromeDriver 자동 관리 (선택적)
fear_and_greed

python-multipart

# Neo4j
neo4j
